+++ akari/pkg/llm/domain/llm.go
@@ -6,7 +6,16 @@ import (
 	"google.golang.org/genai"
 )
 
-// 型エイリアス: 外部ライブラリの型をドメインでラップ.
+type GeminiRepository interface {
+	SendChatMessage(
+		ctx context.Context,
+		systemPrompt string,
+		history []*Content,
+		message string,
+		functions []Function,
+	) ([]*string, []*Part, error)
+}
+
 type (
 	Content          = genai.Content
 	Part             = genai.Part
@@ -14,7 +23,6 @@ type (
 	FunctionResponse = genai.FunctionResponse
 )
 
-// Function はLLMで使用可能な関数を表すドメインモデル.
 type Function struct {
 	FunctionDeclaration *genai.FunctionDeclaration
 	Function            func(ctx context.Context, request *FunctionCall) (map[string]any, error)
